refactor(github): split adapter final files
